backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-unfold==0.37.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.3.0
requests==2.31.0
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==22.0.0
pytest==8.1.1
pytest-django==4.8.0
